Refcator: Changed a little the arhitecture: Moved the CreateQuiz file from the QuizAdmin folder to the Quiz folder
