+++ Duo.Web/Views/Quiz/CreateQuiz.cshtml.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Duo.Web.Views.QuizAdmin
+namespace Duo.Web.Views.Quiz
 {
     public class CreateQuizModel : PageModel
     {
