+++ Duo.Web/Views/QuizAdmin/Index.cshtml
@@ -20,8 +20,8 @@
             </div>
 
             <div class="d-flex gap-2 align-items-center">
-                <a href="@Url.Action("CreateQuiz", "QuizAdmin")" class="btn btn-primary">Add quiz</a>
-                <a href="@Url.Action("ViewQuizzes", "QuizAdmin")" class="btn btn-outline-light btn-sm">
+                <a href="@Url.Action("CreateQuiz", "Quiz")" class="btn btn-primary">Add quiz</a>
+                <a href="@Url.Action("ViewQuizzes", "Quiz")" class="btn btn-outline-light btn-sm">
                     <i class="bi bi-chevron-right fs-4"></i>
                 </a>
             </div>
